requirements.txt
TensorFlowTTS
streamlit==1.10.0
streamlit_option_menu==0.3.2
Soundfile==0.10.3.post1
tensorflow==2.6.0
keras==2.6
pandas
git+https://github.com/repodiac/german_transliterate.git#egg=german_transliterate
playsound
pymodi